Adjusted dependencies for docker container
In theory this should be all that's needed I think, however it still fails on runtime as the f3d application seems to be crashing running in the container when executing the script. I suspect this is because f3d uses OpenGL, which does not support headless rendering (at least not natively).